Splits syndiation access into separate module This makes the code more testable and means the syndication access methods are not so tighly coupled to the init fn
